+++ translations/de.py
@@ -72,6 +72,7 @@ def translateMainDE(window):
     window.btn_riscv32.setText("Durchsuchen") # Browse
     window.btn_riscv64.setText("Durchsuchen") # Browse
     window.pushButton_6.setText("Übernehmen") # Apply
+    window.btn_apply_qemu2.setText("Übernehmen") # Apply
 
     # About tab
     # label_7 = Built on Python and PyQt technology, licensed under GNU General Public License 3.0