+++ translations/fr.py
@@ -72,6 +72,7 @@ def translateMainFR(window):
     window.btn_riscv32.setText("Naviguer") # Browse
     window.btn_riscv64.setText("Naviguer") # Browse
     window.pushButton_6.setText("Appliquer") # Apply
+    window.btn_apply_qemu2.setText("Appliquer") # Apply
 
     # About tab
     # label_7 = Built on Python and PyQt technology, licensed under GNU General Public License 3.0