+++ translations/pt.py
@@ -72,6 +72,7 @@ def translateMainPT(window):
     window.btn_riscv32.setText("Procurar") # Browse
     window.btn_riscv64.setText("Procurar") # Browse
     window.pushButton_6.setText("Aplicar") # Apply
+    window.btn_apply_qemu2.setText("Aplicar") # Apply
 
     # About tab
     # label_7 = Built on Python and PyQt technology, licensed under GNU General Public License 3.0