+++ translations/ru.py
@@ -72,6 +72,7 @@ def translateMainRU(window):
     window.btn_riscv32.setText("Обзор") # Browse
     window.btn_riscv64.setText("Обзор") # Browse
     window.pushButton_6.setText("Применить") # Apply
+    window.btn_apply_qemu2.setText("Применить") # Apply
 
     # About tab
     # label_7 = Built on Python and PyQt technology, licensed under GNU General Public License 3.0