+++ translations/uk.py
@@ -70,6 +70,7 @@ def translateMainUK(window):
     window.btn_riscv32.setText("Огляд") # Browse
     window.btn_riscv64.setText("Огляд") # Browse
     window.pushButton_6.setText("Застосувати") # Apply
+    window.btn_apply_qemu2.setText("Застосувати") # Apply
 
     # About tab
     # label_7 = Built on Python and PyQt technology, licensed under GNU General Public License 3.0